proper epoch conversions

+++ app/src/main/java/me/chaitanyavp/aerozen/RoomActivity.java
@@ -397,20 +397,20 @@ public class RoomActivity extends AppCompatActivity {
     dueDateLayout.setOrientation(LinearLayout.HORIZONTAL);
     dueDateLayout.setLayoutParams(dueDateLayoutParams);
 
-    final HashMap<String,Long> dateTime = new HashMap<>();
-    dateTime.put("year", (long) -1);
-    dateTime.put("month", (long) -1);
-    dateTime.put("day", (long) -1);
-    dateTime.put("hour", (long) -1);
-    dateTime.put("min", (long) -1);
+    final HashMap<String,Integer> dateTime = new HashMap<>();
+    dateTime.put("year", -1);
+    dateTime.put("month", -1);
+    dateTime.put("day", -1);
+    dateTime.put("hour", -1);
+    dateTime.put("min", -1);
 
     final CheckBox dueDateCheckBox = new CheckBox(this);
 
     final TimePickerDialog timePickerDialog = new TimePickerDialog(this, new OnTimeSetListener() {
       @Override
       public void onTimeSet(TimePicker timePicker, int i, int i1) {
-         dateTime.put("hour", (long) i);
-         dateTime.put("minute", (long) i1);
+         dateTime.put("hour", i);
+         dateTime.put("minute", i1);
          dueDateCheckBox.setChecked(true);
       }
     }, 23, 59, false);
@@ -418,9 +418,9 @@ public class RoomActivity extends AppCompatActivity {
     final DatePickerDialog datePickerDialog = new DatePickerDialog(this, new OnDateSetListener() {
       @Override
       public void onDateSet(DatePicker datePicker, int i, int i1, int i2) {
-          dateTime.put("year", (long) i);
-          dateTime.put("month", (long) i1);
-          dateTime.put("day", (long) i2);
+          dateTime.put("year", i);
+          dateTime.put("month", i1);
+          dateTime.put("day", i2);
           timePickerDialog.show();
       }
     }, calendar.get(Calendar.YEAR), calendar.get(Calendar.MONTH), calendar.get(Calendar.DAY_OF_MONTH));
@@ -435,7 +435,7 @@ public class RoomActivity extends AppCompatActivity {
 //                     dueDateCheckBox.setChecked(dateTime.values().contains((long) -1));
                  }
                  else{
-                    dateTime.put("year", (long) -1);
+                    dateTime.put("year", -1);
                  }
              }
          }
@@ -458,11 +458,11 @@ public class RoomActivity extends AppCompatActivity {
       @Override
       public void onClick(DialogInterface dialog, int which) {
           if(task != null){
-
+            //TODO:
           }
           else{
               Task newTask;
-              if(dueDateCheckBox.isChecked() && !dateTime.values().contains((long) -1)){
+              if(dueDateCheckBox.isChecked() && !dateTime.values().contains(-1)){
                   newTask = new Task(userID, taskInput.getText().toString(), priority.getKeyProgressIncrement(), 3);
               }
               else{
